Update message_friends.py : Migrated Message Functions
